requirements.txt
discord
discord.py
youtube-dl
PyNaCl
certifi
charset-normalizer
idna
python-dotenv
requests
urllib3
proxmoxer
PyYAML